apps/ml/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
motor==3.6.0
beanie==1.27.0
celery[redis]==5.4.0
redis==5.2.0
apscheduler==3.10.4
mlflow==2.17.2
aioboto3==13.2.0
httpx==0.27.2
structlog==24.4.0
scikit-learn==1.5.2
numpy>=1.26.4,<2.0
pandas>=2.2.3
matplotlib>=3.9.2
seaborn>=0.13.2
joblib==1.4.2
prometheus-client==0.21.0
python-multipart==0.0.12
aiofiles==24.1.0
cloudpickle==3.1.0
sse-starlette==2.1.3

# HuggingFace / deep learning
transformers==4.47.1
torch==2.5.1
torchvision==0.20.1
accelerate==1.2.1
ultralytics==8.3.52
inference-sdk>=0.9.0
opencv-python-headless==4.10.0.84
Pillow>=10.0.0

# Security
clamd==1.0.2                      # ClamAV daemon client (virus scanning)

# Optional data source dependencies (install as needed)
asyncpg==0.30.0                   # PostgreSQLDataSource
sqlalchemy==2.0.36                # SQLDataSource
aiokafka==0.11.0                  # KafkaDataSource
datasets==3.2.0                   # HuggingFaceDataSource
google-cloud-storage==2.19.0      # GCSDataSource
azure-storage-blob==12.24.0       # AzureBlobDataSource
paramiko==3.5.0                   # FTPDataSource (SFTP)

# Auth
python-jose[cryptography]==3.3.0  # JWT tokens

